Move LU into its own module
